move primary components to primary folder
