+++ client/a8c-for-agencies/sections/reports/primary/overview/index.tsx
@@ -24,7 +24,7 @@ export default function ReportsOverview() {
 			</LayoutTop>
 			<LayoutBody className="a4a-reports-content">
 				<div className="a4a-reports__main-content">
-					<h1>{ translate( 'Client Reports' ) }</h1>
+					<h1>{ title }</h1>
 				</div>
 			</LayoutBody>
 		</Layout>
